Fixing debug print in schedule distribution

+++ network_module/downlink_phase/dynamic_schedule_distribution.cpp
@@ -300,7 +300,7 @@ bool DynamicScheduleDownlinkPhase::recvPkt(long long slotStart, Packet& pkt)
         if(received) {
             AesOcb& ocb = ctx.getKeyManager()->getScheduleDistributionOCB();
             unsigned int tileNumber = ctx.getCurrentTile(slotStart);
-            unsigned int seqNo = 1;
+            unsigned long long seqNo = 1;
             unsigned int masterIndex = ctx.getKeyManager()->getMasterIndex();
             if(ENABLE_CRYPTO_DOWNLINK_DBG)
                 print_dbg("[SD] Verifying downlink: tile=%u, seqNo=%llu, mI=%u\n",
